Warning fixes

+++ src/external/scintilla/include/Platform.h
@@ -193,11 +193,11 @@ public:
 
 	static inline unsigned int ValueOfHex(const char ch) {
 		if (ch >= '0' && ch <= '9')
-			return ch - '0';
+			return (unsigned int)(ch - '0');
 		else if (ch >= 'A' && ch <= 'F')
-			return ch - 'A' + 10;
+			return (unsigned int)(ch - 'A' + 10);
 		else if (ch >= 'a' && ch <= 'f')
-			return ch - 'a' + 10;
+			return (unsigned int)(ch - 'a' + 10);
 		else
 			return 0;
 	}
